Split registerStore call up for readability
