+++ src/audiences/edit/data/index.js
@@ -2,6 +2,7 @@ import {
 	defaultAudience,
 	defaultPost,
 } from './defaults';
+import reducer from './reducer';
 
 const { apiFetch } = wp;
 const { registerStore } = wp.data;
@@ -22,6 +23,12 @@ if ( window.Altis.Analytics.Audiences.Current ) {
 	INITIAL_STATE.posts.push( window.Altis.Analytics.Audiences.Current );
 	INITIAL_STATE.post = window.Altis.Analytics.Audiences.Current;
 }
+
+const controls = {
+	FETCH_FROM_API( action ) {
+		return apiFetch( action.options );
+	},
+};
 
 const actions = {
 	setFields( fields ) {
@@ -57,128 +64,66 @@ const actions = {
 	},
 };
 
-export const store = registerStore( 'audience', {
-	initialState: INITIAL_STATE,
-	reducer( state, action ) {
-		switch ( action.type ) {
-			case 'SET_FIELDS': {
-				return {
-					...state,
-					fields: action.fields,
-				};
-			}
-			case 'ADD_ESTIMATE': {
-				const key = JSON.stringify( action.audience );
-				if ( state.estimates[ key ] ) {
-					return state;
-				}
-				return {
-					...state,
-					estimates: {
-						...state.estimates,
-						[ key ]: action.estimate,
-					},
-				};
-			}
-			case 'ADD_POSTS': {
-				const posts = state.posts.slice();
-				action.posts.forEach( post => {
-					if ( ! posts.filter( existing => post.id === existing.id ).length ) {
-						posts.push( post );
-					}
-				} );
-				if ( state.posts.length === posts.length ) {
-					return state;
-				}
-				return {
-					...state,
-					posts,
-				};
-			}
-			case 'SET_POST': {
-				let posts = state.posts.slice();
-				if ( action.post.id ) {
-					posts = posts.map( post => {
-						if ( post.id === action.post.id ) {
-							post = {
-								...post,
-								...action.post,
-							};
-						}
-						return post;
-					} );
-				}
-				return {
-					...state,
-					posts: action.post.id ? posts : state.posts,
-					post: {
-						...state.post,
-						...action.post,
-					},
-				};
-			}
-			default: {
-				return state;
-			}
+const selectors = {
+	getFields( state ) {
+		return state.fields;
+	},
+	getEstimate( state, audience ) {
+		const key = JSON.stringify( audience );
+		return state.estimates[ key ] || {
+			count: 0,
+			total: 0,
+			histogram: new Array( 28 ).fill( { count: 1 } ), // Build empty histrogram data.
+		};
+	},
+	getPost( state ) {
+		return state.post;
+	},
+	getPosts( state ) {
+		return state.posts;
+	},
+};
+
+const resolvers = {
+	* getFields() {
+		const fields = yield actions.fetch( {
+			path: 'analytics/v1/audiences/fields',
+		} );
+		return actions.setFields( fields );
+	},
+	* getEstimate( audience ) {
+		const audienceQuery = encodeURIComponent( JSON.stringify( audience ) );
+		const estimate = yield actions.fetch( {
+			path: `analytics/v1/audiences/estimate?audience=${ audienceQuery }`,
+		} );
+		return actions.addEstimate( audience, estimate );
+	},
+	* getPost( id ) {
+		const post = yield actions.fetch( {
+			path: `wp/v2/audiences/${ id }?context=edit`,
+		} );
+		if ( post.status === 'auto-draft' ) {
+			post.title.rendered = '';
 		}
+		if ( ! post.audience ) {
+			post.audience = defaultAudience;
+		}
+		yield actions.addPosts( [ post ] );
+		return actions.setPost( post );
 	},
+	* getPosts( page = 1, search = '' ) {
+		const posts = yield actions.fetch( {
+			path: `wp/v2/audiences?context=edit&page=${ page }&search=${ search }`,
+		} );
+		return actions.addPosts( posts );
+	},
+};
+
+export const store = registerStore( 'audience', {
 	actions,
-	selectors: {
-		getFields( state ) {
-			return state.fields;
-		},
-		getEstimate( state, audience ) {
-			const key = JSON.stringify( audience );
-			return state.estimates[ key ] || {
-				count: 0,
-				total: 0,
-				histogram: new Array( 28 ).fill( { count: 1 } ), // Build empty histrogram data.
-			};
-		},
-		getPost( state ) {
-			return state.post;
-		},
-		getPosts( state ) {
-			return state.posts;
-		},
-	},
-	controls: {
-		FETCH_FROM_API( action ) {
-			return apiFetch( action.options );
-		},
-	},
-	resolvers: {
-		* getFields() {
-			const fields = yield actions.fetch( {
-				path: 'analytics/v1/audiences/fields',
-			} );
-			return actions.setFields( fields );
-		},
-		* getEstimate( audience ) {
-			const audienceQuery = encodeURIComponent( JSON.stringify( audience ) );
-			const estimate = yield actions.fetch( {
-				path: `analytics/v1/audiences/estimate?audience=${ audienceQuery }`,
-			} );
-			return actions.addEstimate( audience, estimate );
-		},
-		* getPost( id ) {
-			const post = yield actions.fetch( {
-				path: `wp/v2/audiences/${ id }?context=edit`,
-			} );
-			if ( post.status === 'auto-draft' ) {
-				post.title.rendered = '';
-			}
-			if ( ! post.audience ) {
-				post.audience = defaultAudience;
-			}
-			yield actions.addPosts( [ post ] );
-			return actions.setPost( post );
-		},
-		* getPosts( page = 1, search = '' ) {
-			const posts = yield actions.fetch( {
-				path: `wp/v2/audiences?context=edit&page=${ page }&search=${ search }`,
-			} );
-			return actions.addPosts( posts );
-		},
-	},
+	controls,
+	initialState: INITIAL_STATE,
+	reducer,
+	resolvers,
+	selectors,
 } );
